Converting methods to rake tasks
Instead of relying on a single environment variable (MULTIJOB_KIND) that allows us to select just one kind of test suite to run, we can define the actual test suites in here as independent rake tasks.

This way, they can be orchestrated together into a "MULTIJOB_KIND" type of run (i.e. we don't modify external behaviour), but we maintain the flexibility to invoke independently each of the different rake tasks on the command line, getting rid of some of the complexity in the process.

* the `run_tests_and_handle_reports` has been replaced by the `integrate:parallel` rake task
* the `calculate_tests_to_run`has been replaced by individual rake tasks for each of the runs (e.g. `cucumber_javascript`, `rspec`, `cucumber_non_tagged`, etc.) NOTE: as I do NOT what tests are actually in each of these runs, suggestions on better names there are welcome.
* the `send_test_coverage_to_codeclimate` method is also a new rake task