dnsmasq changes required for installing dependencies
_NOTE: Please, if you're reviewing this PR, do confirm if this is really ok to do._

The installation of dependencies used to happen when the builder docker image was being **built**. And the builder image was being built on ever CI run (in fact, in each of the parallel runs).

I have moved here to a model where the builder image only includes system and not application dependencies. The application dependencies are still installed as part of the CI build (we want this), but this change should make builds somewhat faster / more efficient, given we no longer have to build the rest of the docker image.

Caching of the applications dependencies was already in place, so the speedup will not be _dramatic_ (installing application dependencies was probably the part of the container builder image creation that took most time), but we are still better off not doing the rest of the image building twice.

Anyway, _because_ application dependencies are now being installed inside the builder container, we need some additional entries to dnsmasq, so that internet access is allowed. Here's what was added to dnsmasq:

* rubygems
* github.com
* yarn registry
* npm registry
* nodejs.org
* luarocks (for openresty/apicast dependencies)
* penlight (http://stevedonovan.github.io/)