Separated pipeline in dependencies phase and tests phase
The idea is that rather having `make` start a container where docker.sh installs ALL the dependencies every single time, we will instead have a pipeline where:

1. there is a **parallelized** dependency installation phase: i.e. there is no reason to install dependencies serially - i.e. bundler -> npm -> luarocks. these can be run in parallel.
2. there is a second (parallelized) phase which is JUST for running the tests (and takes installed dependencies as a kind of cache)

To do this, I've extracted the commands that were used to install dependencies from `docker.sh` to `make` targets. This way we can cater for both the `dev` env (contributor would use docker-compose), as well as the `test` / `ci` environment (a bunch of containers would be provisioned for builds to run there natively).

As a benefit of this work, we can now ensure that in our CI builds each of the dependency installation tasks runs ONCE per test suite run (at the moment, they would run for each of the 7 parallel builds, meaning that - even though we don't get any particular speedup due to the fact that it was being executed in parallel - we were still wasting some resources)