[circleci] introduced reusable parameterizable commands
these were introduced in 2.1 and allow us to significantly simplify the build config.

jobs now are much more readable and straightforward to follow, going from a high-level order of steps to a detailed view of each step, either in the yaml anchor or in the circleci command.

I still prefer to use yaml anchors where I don't need parameters, as there is a much greater chance for IDE tools to support YAML than supporting circleci config.