Add mysql2 ld flags instructions in INSTALL.md
Sometimes the mysql2 gem installation fails with an error telling that
the library was not found for linking. This commit adds the instruction
to configure it for bundler.

It also modifies the `Run Porta` instruction to use `env` to set the
`UNICORN_WORKERS` environment variable allowing it to works fine in
fish shell along with bash and zsh.