Include Postgres schema dumping into make schema
After making some change in the database we need to regenerate the
schema for Postgres. This changes the `make schema` command to
accept a env var `POSTGRES_DATABASE_URL` and uses it to migrate
the database and updates the schema after the migration. If not
present, it will use a default
"postgresql://postgres:@localhost:5433/3scale_system_development"
value