Move FilterArguments to app/lib
When we change some code in development and we don't restart the
app it explodes with `A copy of ThreeScale::SidekiqLoggingMiddleware
has been removed from the module tree but is still active!`.

This happens because the ThreeScale::SidekiqLoggingMiddleware class
uses an class that stays inside lib folder.

This commit moves it to app/lib and explicitly adds the ThreeScale
module in the usage.