Remove the usage of require_dependency in Account models
This commit is one of some to move mostly of our lib/* to app/lib
in order to have eager loading on boot

- This one removes the needing of require_dependency
'account/domains'. To this, we needed to rename the
Account::Domains concern to be Account::ProviderDomains in order
to not have naming clashes with the other Domains module.
- Remove the backend_client require_dependency from Account class.