[db] Migrate Database type values
Otherwise it will throw

Handled error
ActionView::Template::Error: The single-table inheritance mechanism failed to locate the subclass: 'AuthenticationProvider::GitHub'.
This error is raised because the column 'type' is reserved for storing the class in case of inheritance.
Please rename this column if you didn't intend it to be used for storing the inheritance class or overwrite AuthenticationProvider.inheritance_column to use another column for that information.