Destroy backend apis belonging to the service when it is deleted
THREESCALE-3317
Currently we have some orphans backend apis in the database caused
by deleting a service. This generates some problems for some
clients when they tries to recreate the service with the same name
and it fails because an already existent backend api with the same
name already exists. This commit changes this behave to delete
all related backend apis when deleting a service when the client
is in the rolling update.

This also adds a rake task to clear orphans backend apis.