Do not create backend_api automatically when create a Service
This commit removes the auto creation of a backend api when a
product is created. We don't need to have a backend api every time
that a product is created.

The idea is that product and backend are two different things that
can be independently created and linked together at will.
All services are/will be products and there will then be no need
for a checkbox.