Fix `Folded services have no overview data` scenario test
This commit fixes this scenario in the cucumber test. Apparently,
it never worked right.

You can test by checking the commit that introduced it and running
```
git checkout 832920cf
features/old/services/multiservice.feature:61
```

It fails because it tries to find an element that does not exist
in the page, because of this, we need to pass the visibile
option so it does not explodes.