Make initial variables available for Edit and Update actions
Currently when something happens when trying to update something
in the integration page, it fails and the registry_policies is
not called because it is only set in the edit action. This commit
changes it and makes it available as a helper method and the view
only calls the method instead of relying on a variable saved in
the action.