Include Rake task to destroy orphans backend apis
Although the deletion process of Services is async, even after
several hours or even days a given Service cannot be created again
using the same system_name from a previously deleted one. This
happens because the backend api is still created. Some customers
reported this bug.

This commit includes a Rake task to allow us to run this in cli
to destroy orphans backend apis