Fix redirect_url field in the proxy and backend_api/mapping rules
We currently have a bug in
app/views/shared/proxy_rules/_form.html.slim page. To render the
redirect_url field, it relies on the owner to check if it is
using proxy pro. But this caller to owner is invalid. Owner for
ProxyRule can be a Proxy or BackendApi, not the Service (which we
expect to call the using_proxy_pro method).