Implement search by pattern for Mapping Rules (#1239)
* Upgrade thinking-sphinx gem to 3.4.2

This commit upgrades thinking sphinx gem. The new version added
the ability to filter using string types, avoinding us to save it
using a CRC32 version of the string.

This commit also upgrades minitest.

* Implement search by pattern for Mapping Rules

This commit implements a search by pattern for mapping rules. It
includes a search field in the mapping rules page for both Backend
APIs and Services. It also adds an index for ProxyRule class using
Sphinx.