Avoid Empty paths for BackendApiConfig
THREESCALE-3631

MySQL and Postgres saves '' as empty string, but Oracle saves it as
null. Now for the first service of each tenant, we link it to a new
backend api, with empty string path in the backend api config. The
problem is that we are checking the uniqueness of service_id and path
for backend_api_config but for Oracle, it searches by empty string
although it is saved as null, so it does not find it by Rails and it
tries to save it but raises an error. We also have this problem if we
try to use where(path: '') or a find_by, that it tries to search by ''
but it is saved as null so it does not find it although it is there.

This commit changes this behave to make the application to save '/'
instead of the empty string.