Fix mapping rules link in Methods and Metrics page
THREESCALE-3719
Currently our Methods and Metrics page have a link to mapping rules
link, but it is linking only to the mapping rules inside the integration
pages and it is not considering the independent mapping rules.

This commit fixes this behave by considering the independent mapping
rules to create the link.