Replace the forked version of HtmlPipeline gem with the released one (#1656)
THREESCALE-3437
We are using a forked version of HtmlPipeline and looks like it isn't
needed, since the fork was only to be able to run with ruby 2.1.

Using this old forked gem is causing us to have a CVE-2018-3740 because
it uses the sanitize library (in a old version). This new version of
HtmlPipeline doesn't even use sanitize.