Fix HTTP BasicAuth when using ProviderKey
THREESCALE-4625
We currently have a bug when apicast tries to fetch the proxy configs
for a service using ProviderKey via BasicAuth. It returns a forbidden
status.

This happened because we changed in #1736 to consider the access_token
when sent in HttpBasicAuth but we only considered if the token was
present.

This commit changes this behave to also consider if the AccessToken is
present, otherwise, it will tries to make the authentication using the
provider key.