Avoid Master account to access routes with ProviderDomain constraint
Currently ProviderDomainConstraint has a bug that it also allows Master
to access it. We should only allow provider on this constraint because
we have another one called MasterOrProviderDomainConstraint to be used
when both can access a route.

This commit changes this behave forbidding MasterDomain to access routes
only for providers.