Fix APIAP routing policy rule
The regex used to match the path of the backend in the APIAP routing policy configuration should only allow paths that _start with_ the chosen value for the path. Currently it matches all paths that _contains_ the value. This fixes it by adding a `^` to the beginning of the pattern.