Remove Test for: Delete BuyerAccount => Delete App & AppKeys in Apisonator.
When a BuyerAccount is destroyed, its Applications/Cinstances are destroyed too.
That will save an ApplicationDeletedEvent for each.
The AppKeys and RefFilters will be saved as destroyed_objects because destroyed_by_association will be a truthy value https://apidock.com/rails/ActiveRecord/AutosaveAssociation/destroyed_by_association .
So this test is redundant now and not so trivial to test as a whole integration/feature anymore.