Prevents new proxy config without backend on proxy update
Prevents that a proxy update generates a new proxy config if the api backend is missing. The api backend can be present as the old legacy `api_backend` attribute of `Proxy` or as a `BackendApi` associated to the service. This issue was particularly troublesome when updating the proxy object via Admin API, where the old `api_backend` attribute is still supported even for APIAP-accounts.