Defaults Paperclip S3 URL style to :s3_domain_url instead of :s3_path_url
This is because we provide two storage options ('filesystem' and 's3'), and the Paperclip `url` option is set in the models to what will actually be the `path` of the attachments, instead of the desired S3 URL style. This used to be OK because Paperclip moves the value set for the `url` option to the `path` option, whenever the storage is 's3'. The problem is that, when Paperclip moves the value from `url` to `path`, it also sets the URL style to `:s3_path_url` (see https://github.com/thoughtbot/paperclip/blob/de92a5a44bae7ed20c33356de5cce584dc9125ae/lib/paperclip/storage/s3.rb#L154-L157), and AWS has deprecated the path style for S3 URLs in favour of the domain style (called "virtual-host style" by Amazon).

So we do not break deployments based on old S3 buckets created before September 30, 2020 (Amazon's cut-off) or integrations with other storage providers compatible with the S3 protocol (e.g. Minio), the override of the default URL style will be avoided and the old default to path style URL reinstated, in case the user opts-in for the `force_path_style` S3 option (https://github.com/3scale/porta/commit/4eb33e94bbc82c5c4963dba252f51844ddd0f7b1).

Closes https://issues.redhat.com/browse/THREESCALE-6955