[gem] Remove sidekiq-pro
Context: we have it used on one of our production deployment.
We mainly rely on 2 features for it:

- Reliability
- Batches

First, after years used by customers on their own deployments without sidekiq-pro, no issues reported to us so far about reliability
We think it is time to say goodbye on this awesome gem.

Second, we successfully make use of an alternative to Pro batches sidekiq-batch

All in one, the necessity of sidekiq-pro dimmed so the decision is to take it down for good