[cucumber] Use meaningful steps
Do not use verbose sentence or steps to describe a behaviour.
Behaviours should be concise and right to the point

Having domain foo.3scale.localhost does not really help

TODO:

Change all "I" to "The user/The administrator"
Be more precise on the roles because "I/You/She/He/We/Us/They"
do not convey the actor role.

Behaviour testing/documentation needs:

- a subject with a role
- an action on an object
- a result

It should be concise and informative enough for any people reading it to understand the feature