async auditing broke rails internals
With Rails 5.1 we relied on the Audit object to be created
and by overriding the internal method `#create_or_update`
to avoid callbacks and it's original logic to be executed
for asynchronous auditing.

Rails 5.2 performs more logic around `#create_or_update`
which causes the described approach to fail.

What we do now is to avoid monkey patching ActiveRecord
methods but override `Audited:Auditor#write_audit`.