use :mem_cache_store
dalli was introduced with
https://github.com/3scale/system/commit/5f59858f4ff

In dalli 2.7 MD5 checksum is hardcoded and we need
to get rid of that. But in 3.0 :dalli_store was
removed. Unfortunatelu 3.0 also does not support
Ruby 2.4

So following official recommendation to use the
Rails built-in cache store implementation still
with Dalli 2.7 but monkey patch key truncation.