mass-assign warning with Invoice creation_time
`Invoice#creation_time` is not good to be under
`attr_accessible` because we don't want users to change it.

Issue was introduced in 3scale/system@2d0bcdd7cf8
which caused warnings when assigning it.

This has to show we better move to approving attributes in
a modern way instead of using `protected_attributes`.