Remove Timecop.scale from the specs
The context for the PUT api endpoints asserts that the given resource is in fact updated. It does so by checking that resource.updated_at is not equal to resource.created_at. When the test for the PUT endpoint is executed less than one second after the test for the POST endpoint, the updated_at attribute is not updated because the rails timestamps have a precision of 1 second. Timecop.scale was being used to accelerate the time during the test execution to ensure the value in updated_at changes in all cases. Now this has been replaced by manually set created_at and updated_at to one second ago, before running the test.