THREESCALE-8027: Replace Timecop with ActiveSupport::Testing::TimeHelpers (#3062)
* Replace Timecop.freeze, Timecop.return and Timecop.travel by Rails TimeHelpers

* Remove Timecop.scale from the specs

The context for the PUT api endpoints asserts that the given resource is in fact updated. It does so by checking that resource.updated_at is not equal to resource.created_at. When the test for the PUT endpoint is executed less than one second after the test for the POST endpoint, the updated_at attribute is not updated because the rails timestamps have a precision of 1 second. Timecop.scale was being used to accelerate the time during the test execution to ensure the value in updated_at changes in all cases. Now this has been replaced by manually set created_at and updated_at to one second ago, before running the test.

* Remove Timecop.scale from the cucumber config

It seems to not have any effect

* Remote timecop gem

* Fix broken tests

Were broken by an unattended find + replace

* TimeHelpers: avoid nested time travel on tests

* Remove Time.freeze calls

The author probably wanted to call Timecop.freeze

* Properly include TimeHelpers in Minitest

* Avoid nested time travel in features

* Rollback timeco_helpers

Renamed as time_machine_helpers

* Time helpers steps: Modify steps to match the old Timecop behaviour

* Fix some cucumber tests after removing Timecop

* Revert "Time helpers steps: Modify steps to match the old Timecop behaviour"

This reverts commit d824137849f4d8dc8dd0c75aaf43e17625588ea8.

* Add a comment to explain the Timecop.scale() problem

* Add # frozen_string_literal: true to new classes

* THREESCALE-8027: Remove superfluous comment

* THREESCALE-8027: Remove whitespace in function calls

To follow the project style

* THREESCALE-8027: Remove unneeded time travel

* THREESCALE-8027: Remove another superfluous comment

* THREESCALE-8027: USe hooks to avoid nested time travel

* THREESCALE-8027: Write a more understandable comment

* THREESCALE-8027: Use blocks on tests when possible

* THREESCALE-8027: Avoid nested time travel in bill tests

In BillVariableForPlanChangedTest

* THREESCALE-8027: Remove unneeded time resets

* THREESCALE-8027: Add a teardown to reset time for user tests

* THREESCALE-8027: Remove references to removed method no_nested_travel_to

* THREESCALE-8027: Implement safe nested traveling

* THREESCALE-8027: Fix applications cucumbers

* THREESCALE-8027: Applications steps: use modern dictionary syntax

For the new step, older steps are not modified

* THREESCALE-8027: Use blocks in applications unit tests

* THREESCALE-8027: Remove unneeded teardown in bill tests

* THREESCALE-8027: Remove unneeded teardown tests configuration

* THREESCALE-8027: Ensure restoring previous time for nested travel

* THREESCALE-8027: Fix codeclimate claims

* THREESCALE-8027: Remove redundant callbacks to restore time

* THREESCALE-8027: Fix PR commits