fix rake in production environment
issue introduced with a53c459

Because `license_finder` is a development dependency and these are
not installed in images, the `require` statement should only be
evaluated if one tries to call the dependent tasks.

Otherwise any rake call raises an error to load `license_finder`.

This is consistent with how it worked when we used cli.