Update thinking_sphinx.yml configuration file:
- add batch_size configuration (see below)
- remove the environments development, test, preview that are not needed in the container image
- group the different settings according to what they configure (thinking sphinx, searchd, etc.)

batch_size setting is required for reindexing in SaaS, as with the default value 1000 reindexing with `rake ts:rt:index` fails with the following error:
```
ThinkingSphinx::QueryLengthError: The supplied SphinxQL statement is 10266420 characters long. The maximum allowed length is 8388603.
```
Probably that's because of indexing CMS contents that can be quite long.