Zeitwerk Preparation: Reorganize constants names and directory tree
:classic and :zeitwerk autoload find the constants in a different way:

 The old :classic autoloader starts from a missing constant: e.g. API.
   1. It runs "API".underscore and the result is "api".
   2. It looks for a file or folder called api and expects the constant to be defined there.

 The new autoloader :zeitwerk starts from a file name, e.g. api.rb.
   1. It runs "api".camelize and the result is "Api".
   2. It expects the constant Api to be defined there.

Most of times, "ConstantName".underscore.camelize == "ConstantName", but that's not always the case, like in the example above.

In this commit, some constants are renamed/moved to get a file strcuture which is good for both :classic and :zeitwerk.
On the other hand, some monkey patches were interfering with the autoload path and causing some issues in zeitwerk, those have been moved out of the autoload path.

This new strcture should work in rails 6.

More info: https://guides.rubyonrails.org/classic_to_zeitwerk_howto.html