Fix deprecation of ActiveMOdel::Error enumeration
Liquid::Filters::FormHelpersTest#test_inline_errors DEPRECATION WARNING: Enumerating ActiveModel::Errors as a hash has been deprecated.
In Rails 6.1, `errors` is an array of Error objects,
therefore it should be accessed by a block with a single block
parameter like this:

person.errors.each do |error|
  attribute = error.attribute
  message = error.message
end

You are passing a block expecting two parameters,
so the old hash behavior is simulated. As this is deprecated,
this will result in an ArgumentError in Rails 7.0.
 (called from each at /opt/ci/workdir/lib/developer_portal/lib/liquid/drops/errors.rb:48)