postgres more convenient test db url
To run with Postgres locally is a little harder because when
DATABASE_URL is set, it remains the same for all environments.

To differentiate from the development database one can set
TEST_ENV_NUMBER but that makes database name like
`3scale_system_development_test` instead of simpler `3scale_system_test`
as on MySQL/MariaDB and more importantly this is somehow ugly to see.

With this change one can just set DATABASE_URL for development and URL
for the `test` environment will be adjusted automatically.

This is also similar to how oracle works where `test` is appended to
`systempdb` automatically.