THREESCALE-11931 fix ThinkingSphinx::QueryError
Sometimes after start-up Sidekiq pods produce ThinkingSphinx::QueryError

> unknown column: 'sphinx_internal_class_name' - REPLACE INTO account_core (id, `sphinx_internal_class_name`, `name`, `account_id`, `username`, `user_full_name`, `email`, `user_key`, `app_id`, `app_name`, `user_id`, `sphinx_internal_id`, `sphinx_internal_class`, `sphinx_deleted`, `sphinx_updated_at`, `provider_account_id`, `tenant_id`, `state`) VALUES (14978, 'Account', 'org_name-lkdjfggf-serch-yyith8w', '156', 'username-lkdjfggf-serch-xskw25c', ' ', 'username-lkdjfggf-serch-xskw25c@anything.invalid', 'af61253a5933dd9ba30bebb6661d16d1 ede2926ef59ecdc45da4a8960164ffc2', '3173be1c fc507eb9', 'org_name-lkdjfggf-serch-yyith8w's App ui_accou-lkdjfggf-accont-vmt8wba', '180', 156, 'Account', 0, 1697580778, 2, 2, 'approved')

A potential cause for this is load order of thinking-sphinx indices vs the monkey patch we are removing here.

Since we use thinking-sphinx 5.5 which has this logic alread merged in pat/thinking-sphinx#1249 , we can just remove the monkey patch and a race condition should not have any opportunit to happen.