fix oracle upgrades from 2.15
The issue was that older version used long identifier names but on Rails
upgrade to 7.0 we falsely believed that our monkey patch for longer
identifiers was not needed anymore (see 939e15a3).

When it fact it was needed when creating table primary keys
auto-increments.

This resulted in `OCIError: ORA-02289: sequence does not exist` errors
when trying to create a new ProxyConfigAffectingCchanges object because
previously the full sequence name was used
PROXY_CONFIG_AFFECTING_CHANGES_SEQ and after upgrade a truncated PROXY_CONFIG_AFFECTING_CHA_SEQ was used.

Now 3scale after upgrades will work.

An issue could be new installations where the sequence was created with
3scale 2.16 resulting in the truncated sequence name but after this fix,
it will try to use the longer sequemce name.

This shouldn't have happened though due to oracle not being supported
with any affected version.